requirements.txt
Django~=5.1.3
PyMySQL~=1.1.1
python-decouple~=3.8
qrcode~=8.0
xhtml2pdf~=0.2.16